+++ ContentPatcher/Framework/Patches/IncludePatch.cs
@@ -83,7 +83,7 @@ namespace ContentPatcher.Framework.Patches
                     // validate file existence
                     if (!this.FromAssetExists())
                     {
-                        this.Monitor.Log($"{errorPrefix}: file doesn't exist.");
+                        this.Monitor.Log($"{errorPrefix}: file '{this.FromAsset}' doesn't exist.");
                         return true;
                     }
 
@@ -109,6 +109,11 @@ namespace ContentPatcher.Framework.Patches
 
                     // load raw file
                     var content = this.ContentPack.LoadAsset<ContentConfig>(this.FromAsset);
+                    if (content?.Changes == null)
+                    {
+                        this.Monitor.Log($"{errorPrefix}: file '{this.FromAsset}' doesn't contain a changes field. Is the file formatted correctly?", LogLevel.Warn);
+                        return true;
+                    }
 
                     // validate fields
                     string[] invalidFields = this.GetInvalidFields(content).ToArray();